requirements.txt
aiodns==3.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
ast-comments==1.2.2
attrs==24.2.0
cachetools==5.5.0
ccxt==4.4.33
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
datasieve==0.1.7
dateparser==1.2.0
fastapi==0.115.5
freqtrade==2024.10
freqtrade-client==2024.10
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
humanize==4.11.0
idna==3.10
janus==1.1.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lightgbm==4.5.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
numpy==1.26.4
orjson==3.10.11
packaging==24.2
pandas==2.2.3
pandas_ta==0.3.14b0
prompt-toolkit==3.0.36
propcache==0.2.0
psutil==6.1.0
py_find_1st==1.1.6
pyarrow==18.0.0
pycares==4.4.0
pycoingecko==3.2.0
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.0
pydantic_core==2.27.0
Pygments==2.18.0
PyJWT==2.10.0
python-binance==1.0.22
python-dateutil==2.9.0.post0
python-rapidjson==1.20
python-telegram-bot==21.7
pytz==2024.2
questionary==2.0.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
schedule==1.2.2
scikit-learn==1.5.2
scipy==1.14.1
sdnotify==0.3.2
setuptools==75.6.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
ta-lib==0.5.1
tabulate==0.9.0
technical==1.4.4
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.1
wcwidth==0.2.13
websockets==14.1
yarl==1.17.2
